+++ changelog.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Mockable members that are considered obsolete will prevent a mock from being created (issue [#240](https://github.com/JasonBock/Rocks/issues/240))
 - Methods that hide mockable methods are no longer generated (issue [#239](https://github.com/JasonBock/Rocks/issues/239))
 - Addressed when explicit interface implementations should occur in shims (issue [#231](https://github.com/JasonBock/Rocks/issues/231))
+- Correctly detecting when duplicate method signatures exist on the same type (issue [#232](https://github.com/JasonBock/Rocks/issues/232))
 
 ## [7.1.1] - 2023-07-23