+++ changelog.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Correctly detecting when duplicate method signatures exist on the same type (issue [#232](https://github.com/JasonBock/Rocks/issues/232))
 - Abstract members that match an existing non-virtual member will prevent a mock from being made (issue [#247](https://github.com/JasonBock/Rocks/issues/247))
 - Added the null forgiving operator for shim members (issue [#246](https://github.com/JasonBock/Rocks/issues/246))
+- Updated type generation for `Task` types in makes to use a FQN (issue [#248](https://github.com/JasonBock/Rocks/issues/248))
 
 ## [7.1.1] - 2023-07-23