+++ changelog.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [7.1.3] - Not Yet Released
+
+### Fixed
+
+- Duplicate constructors will prevent a type from being mocked (issue [#256](https://github.com/JasonBock/Rocks/issues/256))
+
 ## [7.1.2] - 2023-07-26
 
 ### Fixed